Flatten lib module hierarchy
Submodules currently only created unnecessary complexity. It should be
easy to explore source code.
